requirements.txt
streamlit
google-generativeai
gspread